Extract favorite toggle button to own component
