+++ superhero-directory/src/features/toggle-favorite/index.ts
@@ -1,0 +1,1 @@
+export { ToggleFavorite } from './toggle-favorite';
